[COOK-2322] Fix Debian init.d template
COOK-2254 left runit service broken on Debian. It was passing `:options`
hash to the template that was still expecting `:params`. Just passing
the `new_resource.options` is not a correct solution as that was meant
for runit templates and is empty by default.

This commit fixes the issue by passing and using only the service name
to the template. The runit commands are hardcoded the same way as they
already are in the provider.