Adjust recipe to optionally use prebuilt RPM on EL
On Enterprise Linux systems, optionally specify that we should install
the runit package via yum instead of building the RPM. This is for users
who already have a Yum repository with the runit RPM available. We add
the attribute "use_package_from_yum" to control this behavior. To
preserve backwards compatibility the attribute defaults to "false".