Skip runit installation if already installed
If there is an executable at the desired installation location, we can
assume that runit is already installed. Unfortunately, I can't find a
way to verify that the *correct version* is installed, but given that
the version isn't configurable via attributes, it seems reasonable that
when/if there ever is a new version of runit, this logic can be changed.