#60 allow waiting for service files in docker too
If in docker and if the `runsvdir` process is running, we can wait for
the service directory to be initialized by `runsvdir.` (To avoid the
error: `warning: /etc/service/{service_name}: unable to open
supervise/ok: file does not exist`).

It is not enough to check only `supervise/ok` and `log/supervise/ok`
files, we also need to wait for `runsv` process. (To avoid the error
`fail: {service_name}: runsv not running`).