add manage_env_dir attribute to runit_service
Per #144, deleting any files in the environment directory not created by the
runit_service instance breaks other cookbooks that rely on
runit_service and manage the contents of the env dir "out of band".
This change restores the default behavior prior to the 1.7 release,
adding a `manage_env_dir` attribute with a default value of
`false`. Setting `manage_env_dir` to `true` activates the behavior to
delete any files in the env dir which were not passed as part of the
runit_service's `env` hash.