Revert `manage_env_dir` feature, don't clean up env dir when env is empty
In preparing changelog notes for a new release I realized that we are adding
a feature flag to block cleaning up the environment, when this might be better
treated as a regression from the [behavior introduced in
v1.6.0](https://github.com/hw-cookbooks/runit/blob/v1.6.0/libraries/provider_runit_service.rb#L108).

In the prior implementation, "zapping" env files which aren't explicitly
declared in Chef is only called if the Chef runit_service resource has a
non-empty `env` hash. The changes in #107 reimplemented the "zapping"
behavior, but did not retain the same guard preventing action when
`env` is empty. This change should resolve that regression.