Enable runit installation in OEL systems
The `imeyer/runit` packagecloud repo does not have the ability to
resolve packages for OEL systems.

By consuming [PR#32](https://github.com/computology/packagecloud-cookbook/pull/32) this
commit enables `runit` cookbook to be able to install in OEL
systems by forcing the os platform to be `rhel` since that is the base
and packages should work the same way.

Pointing to packagecloud repo (master) in github since there is where
the new code for forcing exist. (Need to release it to Supermarket)