Mark env files as sensitive (#182)
* Mark env files as sensitive

This supresses env file logging output in case of sensitive credentials making their way in to that directory. There is a fallback in case this is an older version of Chef that does not support sensitive as a common resource argument.

* Use Chef::Resource instead for feature detection

In Chef 11.10 it can come up with a false positive so using just Resource will work as expected instead.

* Make sure env files are handled sensitively

* Use create_file resource instead to test parameter

The with matcher is not available for the render_file resource.