Restore limited SUSE support
We yanked SUSE support a while back, because there weren't packages
for runit. However Chef Server, which provides it's own copy of runit,
uses this cookbook to manage its install on SUSE.

Testing this is problematic, in that I don't have a package to install
in the kitchen test environment.

Signed-off-by: Mark Anderson <mark@chef.io>