Use the resource DSL in the resource
We converted the service resource in Chef 15 (currently master) to use the DSL instead of doing everything as instance variables. This broke this resource as it was doing some less than best practice things. With this change we're using the DSL while still inheriting from the service resource class. It's still a HWRP, but it's a less crazy HWRP and one that's less likely to break as we refactor code throughout Chef.

Signed-off-by: Tim Smith <tsmith@chef.io>