move idiocy to after_created block
after writing this i think this is actually correct because we're
accessing values off of "new_resource" that might be set by the user
so we have to do this in an after_created block after the new_resource
has accumulated all its properties.

trickery is because there is no "new_resource" off of the resource,
and because we're inside of a block that is evaluated against the
service resource we're constructing, so we need a copy of "self" when
it was the runit_service resource.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>