Add a new use_init_script_sv_link property
I'm not adding this to the readme as this is something that should
*really* only be used by Chef internally. Basically our issue is that
that runit cookbook now drops an init script on Ubntu hosts instead of a
link to sv. The link to sv combined with our control file would allow us
to properly signal services with their full path. When you use the init
script that gets laid down you only get the short name like 'postgresl'
which won't help when you're in an omnibus package.1

Signed-off-by: Tim Smith <tsmith@chef.io>