wait until runsv is reading from supervise/ok
If the supervise/ok pipe already exists on disk before runsv has
started, then existence check here is not enough.

Since runsv opens the pipe for reading as part of its startup, we can
open it for writing which will block until runsv has called open.

A version of this fix has been running in one of our internal projects
for some time.

Signed-off-by: Steven Danna <steve@chef.io>